+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
+- **Changed:** Optimized `BaseField` props calculation.
 - **Fixed:** `Bridge.check` error message.
 
 ## [v1.18.0](https://github.com/vazco/uniforms/tree/v1.18.0) (2017-04-30)